Consolidated all the code pertaining to the player to one file and made it so the player will not stick to walls
