+++ Assets/Scripts/capsuleMovement.cs
@@ -1,0 +1,48 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class capsuleMovement : MonoBehaviour
+{
+    public float speed = 2f;
+    public float jump = 200f;
+    public float wallCheckDistance = 0.001f;
+
+    private float move;
+    private Rigidbody2D rb;
+
+    void Start()
+    {
+        rb = GetComponent<Rigidbody2D>();
+    }
+
+    void Update()
+    {
+        Move();
+        Jump();
+    }
+
+    void Move()
+    {
+        float horizontalInput = Input.GetAxis("Horizontal");
+
+        rb.velocity = new Vector2(horizontalInput * speed, rb.velocity.y);
+    }
+
+    void Jump()
+    {
+        if (Input.GetButtonDown("Jump"))
+        {
+            rb.AddForce(new Vector2(rb.velocity.x, jump));
+        }
+    }
+
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.transform.tag == "Lava")
+        {
+            SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
+        }
+    }
+}
