--- Assets/capsuleJump.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c5e2687f435d9194a97d03e8ecf665ce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
