--- Assets/lavaCollide.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class lavaCollide : MonoBehaviour
-{
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        if(collision.transform.tag == "Lava")
-        {
-            SceneManager.LoadScene("Scene2");
-        }
-    }
-}
