--- Assets/lavaCollide.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3eea72935c21c744295b6c7d88c31a74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
